refactor : Booth Entity 필드 수정 및 파일 위치 변경
1. 물리 모델링에 따른 필드 정보 변경
- 삭제여부를 위한 is_deleted 필드 추가
2. 도메인 디렉토리 구조로 파일 위치 변경
