+++ src/main/java/com/hallym/festival/domain/booth/controller/BoothController.java
@@ -1,7 +1,8 @@
-package com.hallym.festival.domain.booth;
+package com.hallym.festival.domain.booth.controller;
 
+import com.hallym.festival.domain.booth.entity.Booth;
+import com.hallym.festival.domain.booth.service.BoothService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -11,7 +12,6 @@ import java.util.List;
 @RestController
 @RequestMapping("/booths")
 @RequiredArgsConstructor
-@Log4j2
 public class BoothController {
 
     private final BoothService boothService;
