+++ src/main/java/com/hallym/festival/domain/booth/entity/BoothType.java
@@ -1,0 +1,5 @@
+package com.hallym.festival.domain.booth.entity;
+
+public enum BoothType {
+    주점, 푸드트럭, 부스, 플리마켓
+}
