+++ src/main/java/com/hallym/festival/domain/booth/repository/BoothRepository.java
@@ -1,5 +1,6 @@
-package com.hallym.festival.domain.booth;
+package com.hallym.festival.domain.booth.repository;
 
+import com.hallym.festival.domain.booth.entity.Booth;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
